+++ src/components/ValidatorsModal/components/CollapseProperties/hook.ts
